Update utils for consistency

+++ src/utils.cpp
@@ -20,7 +20,7 @@ bool read_temparatures(double &cpu_temp, double &gpu_temp){
     }
     
     cpu_temp = (double)atoi(fgets(cpu, 6, fcputemp))/1000;
-    gpu_temp = atoi(fgets(gpu, 6, fgputemp))/1000;
+    gpu_temp = (double)atoi(fgets(gpu, 6, fgputemp))/1000;
     
     printf("\rCpu : %.2fC, Gpu : %.2fC.", cpu_temp, gpu_temp);
